Création d'une navbar pour les voitures et composants d'exemple pour un développement plus fluide

+++ angularwithasp.client/src/app/components/car-list/car-list.component.ts
@@ -1,31 +1,42 @@
 import { Component, OnInit } from '@angular/core';
 import { CarService } from '../../services/car.service';
 import { Car } from '../../models/car.model';
+import { CommonModule } from '@angular/common';
+import { Router } from '@angular/router';
 
 @Component({
-    selector: 'app-car-list',
-    standalone: false,
-    templateUrl: './car-list.component.html',
-    styleUrls: ['./car-list.component.scss']
+  selector: 'app-car-list',
+  templateUrl: './car-list.component.html',
+  styleUrls: ['./car-list.component.scss'],
+  imports: [CommonModule]
 })
 export class CarListComponent implements OnInit {
-    cars: Car[] = [];
+  cars: Car[] = [];
 
-    constructor(private carService: CarService) { }
+  constructor(
+    private carService: CarService,
+    private router: Router
+  ) { }
 
-    ngOnInit(): void {
-        this.loadCars();
-    }
+  ngOnInit(): void {
+    this.loadCars();
+  }
 
-    loadCars(): void {
-        this.carService.getCars().subscribe(cars => {
-            this.cars = cars;
-        });
-    }
+  loadCars(): void {
+    this.carService.getCars().subscribe(cars => {
+      this.cars = cars;
+    });
+  }
 
     deleteCar(id: number): void {
-        this.carService.deleteCar(id).subscribe(() => {
-            this.loadCars();
-        });
+        if (confirm('Etes vous sur de vouloir supprimer ce véhicule?')) {
+            this.carService.deleteCar(id).subscribe(() => {
+                this.loadCars();
+            });
+        }
     }
+
+  updateCar(id: number): void {
+    this.router.navigate(['car-edit', id]);
+  }
 }
